+++ CHANGELOG.md
@@ -4,11 +4,28 @@ Changelog
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+
 Unreleased
+***********
+
+
+4.0.1
+***********
+* Updates email, pdfs endpoints for the PHT to use Role.Any
+
+4.0.0
 ***********
 * Fixes the issue when a technical review already exists so a new one is not created
 * [BREAKING] Updates the PHT endpoints such that they are coherent e,g `POST entity/create`
 * [BREAKING] Updates the retrieval of entities endpoint from `entity/list/{id}` to be `entity/users/{id}/entity`
+* Panel `auto-create` can now update the submitted proposals for `Science Verification`
+* Fixes the bug with the report endpoint.
+* Creates Science reviews as part of the panel update endpoint.
+* Updates ODA version to 11.5.0
+* Add `/odt/prsls/project-view` endpoint which returns a view for the UI of an outer join of the Proposals and Projects
+* Write user from auth context to entity metadata for ODT endpoints
+* [BREAKING] Use Helm global values for ODA configuration rather than ska-oso-services chart ones
+* Deploy Secret for ODA password rather than relying on external one from ODA chart
 
 3.1.1
 ***********